+++ Source/Objects/Data Processing/Histogram Objects/Waveform/ORWaveformController.m
@@ -329,10 +329,10 @@
 - (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row
 {
 	if([[tableColumn identifier] isEqualToString:@"Value"]){
-		return [NSNumber numberWithInteger:[model value:row]];
+		return [NSNumber numberWithInteger:[model value:(uint32_t)row]];
 	}
     else if([[tableColumn identifier] isEqualToString:@"Hex"]){
-		return [NSString stringWithFormat:@"0x%x",[model value:row]];
+		return [NSString stringWithFormat:@"0x%x",[model value:(uint32_t)row]];
 	}
 
 	else return [NSNumber numberWithInteger:row];
